Don't try to chown on windows.
