+++ core/paths/paths.go
@@ -6,10 +6,7 @@ package paths
 
 import (
 	"os"
-	"os/user"
-	"strconv"
 
-	"github.com/juju/errors"
 	jujuos "github.com/juju/os"
 	"github.com/juju/os/series"
 )
@@ -32,9 +29,6 @@ const (
 	instanceCloudInitDir
 	cloudInitCfgDir
 	curtinInstallConfig
-
-	// LogfilePermission is the file mode to use for log files.
-	LogfilePermission = os.FileMode(0640)
 )
 
 const (
@@ -77,6 +71,10 @@ var winVals = map[osVarType]string{
 	metricsSpoolDir:  "C:/Juju/lib/juju/metricspool",
 	uniterStateDir:   "C:/Juju/lib/juju/uniter/state",
 }
+
+// Chown is a variable here so it can be mocked out in tests to a no-op.
+// Agents run as root, but users don't.
+var Chown = os.Chown
 
 // osVal will lookup the value of the key valname
 // in the appropriate map, based on the series. This will
@@ -185,62 +183,4 @@ func MustSucceed(s string, e error) string {
 		panic(e)
 	}
 	return s
-}
-
-// SetSyslogOwner sets the owner and group of the file to be the appropriate
-// syslog users as defined by the SyslogUserGroup method.
-func SetSyslogOwner(filename string) error {
-	user, group := SyslogUserGroup()
-	return SetOwnership(filename, user, group)
-}
-
-// Chown is a variable here so it can be mocked out in tests to a no-op.
-// Agents run as root, but users don't.
-var Chown = os.Chown
-
-// SetOwnership sets the ownership of a given file from a path.
-// Searches for the corresponding id's from user, group and uses them to chown.
-func SetOwnership(filePath string, wantedUser string, wantedGroup string) error {
-	group, err := user.LookupGroup(wantedGroup)
-	if err != nil {
-		return errors.Trace(err)
-	}
-	gid, err := strconv.Atoi(group.Gid)
-	if err != nil {
-		return errors.Trace(err)
-	}
-	usr, err := user.Lookup(wantedUser)
-	if err != nil {
-		return errors.Trace(err)
-	}
-	uid, err := strconv.Atoi(usr.Uid)
-	if err != nil {
-		return errors.Trace(err)
-	}
-	return Chown(filePath, uid, gid)
-}
-
-// PrimeLogFile ensures that the given log file is created with the
-// correct mode and ownership.
-func PrimeLogFile(path string) error {
-	f, err := os.OpenFile(path, os.O_CREATE|os.O_WRONLY, LogfilePermission)
-	if err != nil {
-		return errors.Trace(err)
-	}
-	if err := f.Close(); err != nil {
-		return errors.Trace(err)
-	}
-	return SetSyslogOwner(path)
-}
-
-// SyslogUserGroup returns the names of the user and group that own the log files.
-func SyslogUserGroup() (string, string) {
-	switch jujuos.HostOS() {
-	case jujuos.CentOS:
-		return "root", "adm"
-	case jujuos.OpenSUSE:
-		return "root", "root"
-	default:
-		return "syslog", "adm"
-	}
 }
